database/requirements.txt
supabase==2.7.4
python-dotenv==1.0.0
psycopg2-binary==2.9.9
